TFP-6109: rydd i utils for fakta-arbeid-og-inntekt
