+++ packages/fakta/arbeid-og-inntekt/src/components/ArbeidOgInntektFaktaPanel.tsx
@@ -12,10 +12,8 @@ import {
 } from '@navikt/fp-kodeverk';
 import { SettPaVentModalIndex } from '@navikt/fp-modal-sett-pa-vent';
 import type {
-  AoIArbeidsforhold,
   ArbeidOgInntektsmelding,
   ArbeidsgiverOpplysningerPerId,
-  Inntektsmelding,
   ManglendeInntektsmeldingVurdering,
   ManueltArbeidsforhold,
 } from '@navikt/fp-types';
@@ -23,128 +21,12 @@ import type { FaktaAksjonspunkt } from '@navikt/fp-types-avklar-aksjonspunkter';
 import { useMellomlagretFormData, usePanelDataContext } from '@navikt/fp-utils';
 
 import { useIsFormDirty } from '../DirtyFormProvider';
-import type { ArbeidsforholdOgInntektRadData, Avklaring } from '../types/arbeidsforholdOgInntekt';
-import { harMatchendeArbeidsgiverIdent, lagArbeidsgiver } from '../utils/arbeidsgiverUtils';
-import { finnInntektsmeldingerForArbeidsgiver } from '../utils/inntektsmeldingUtils';
+import type { ArbeidsforholdOgInntektRadData } from '../types/arbeidsforholdOgInntekt';
+import { byggTabellStruktur, finnUløstArbeidsforholdIndex } from '../utils/arbeidOgInntektTabellUtils.ts';
 import { ArbeidsforholdRad } from './ArbeidsforholdRad';
 import { ArbeidsOgInntektOverstyrPanel } from './ArbeidsOgInntektOverstyrPanel';
 
 import styles from './arbeidOgInntektFaktaPanel.module.css';
-
-// TODO(siri): flytt hjelpefuksjoner til bunn eller til egne filer
-const sorterTabell = (radX: ArbeidsforholdOgInntektRadData, radY: ArbeidsforholdOgInntektRadData): number => {
-  const radXHarAp = radX.årsak;
-  const radYHarAp = radY.årsak;
-  if (radXHarAp && !radYHarAp) {
-    return -1;
-  }
-  if (radYHarAp && !radXHarAp) {
-    return 1;
-  }
-  return radX.arbeidsgiverNavn.localeCompare(radY.arbeidsgiverNavn);
-};
-
-const lagAvklaringFraArbeidsforhold = (
-  arbeidsforhold: AoIArbeidsforhold,
-  arbeidsgiverNavn: string,
-): Avklaring | undefined => {
-  const { fom, tom, saksbehandlersVurdering, stillingsprosent, begrunnelse } = arbeidsforhold;
-  if (!arbeidsforhold.saksbehandlersVurdering) {
-    return undefined;
-  }
-  if (
-    arbeidsforhold.saksbehandlersVurdering === ArbeidsforholdKomplettVurderingType.MANUELT_OPPRETTET_AV_SAKSBEHANDLER ||
-    arbeidsforhold.saksbehandlersVurdering === ArbeidsforholdKomplettVurderingType.OPPRETT_BASERT_PÅ_INNTEKTSMELDING
-  ) {
-    return {
-      arbeidsgiverNavn,
-      fom,
-      tom,
-      stillingsprosent,
-      saksbehandlersVurdering: saksbehandlersVurdering ?? undefined,
-      begrunnelse: begrunnelse ?? undefined,
-    };
-  }
-  return {
-    saksbehandlersVurdering: saksbehandlersVurdering ?? undefined,
-    begrunnelse: begrunnelse ?? undefined,
-  };
-};
-
-const lagAvklaringFraInntektsmelding = ({
-  saksbehandlersVurdering,
-  begrunnelse,
-}: Inntektsmelding): Avklaring | undefined => {
-  return saksbehandlersVurdering
-    ? {
-        saksbehandlersVurdering,
-        begrunnelse: begrunnelse ?? undefined,
-      }
-    : undefined;
-};
-
-const sorterGittÅrsak = (arbeidsforhold1: AoIArbeidsforhold) => (arbeidsforhold1.årsak ? -1 : 1);
-
-const byggTabellStruktur = (
-  arbeidOgInntekt: ArbeidOgInntektsmelding,
-  arbeidsgiverOpplysningerPerId: ArbeidsgiverOpplysningerPerId,
-): ArbeidsforholdOgInntektRadData[] => {
-  const { arbeidsforhold, inntektsmeldinger, inntekter } = arbeidOgInntekt;
-
-  const alleArbeidsforhold = [...arbeidsforhold.sort(sorterGittÅrsak)].reduce<ArbeidsforholdOgInntektRadData[]>(
-    (acc, af) => {
-      const tidligereAf = acc.find(harMatchendeArbeidsgiverIdent(af));
-      if (tidligereAf) {
-        return acc;
-      }
-      const arbeidsforholdForRad = arbeidsforhold.filter(harMatchendeArbeidsgiverIdent(af));
-      const arbeidsgiverOpplysninger = lagArbeidsgiver(
-        af.arbeidsgiverIdent,
-        arbeidsgiverOpplysningerPerId[af.arbeidsgiverIdent],
-      );
-      const inntektsmeldingerForRad = finnInntektsmeldingerForArbeidsgiver(inntektsmeldinger, af.arbeidsgiverIdent);
-      const inntektsposter = inntekter.find(harMatchendeArbeidsgiverIdent(af))?.inntekter ?? [];
-
-      const ne: ArbeidsforholdOgInntektRadData = {
-        ...arbeidsgiverOpplysninger,
-        årsak: af.årsak ?? inntektsmeldingerForRad[0]?.årsak ?? undefined,
-        avklaring: lagAvklaringFraArbeidsforhold(af, arbeidsgiverOpplysninger.arbeidsgiverNavn),
-        arbeidsforholdForRad,
-        inntektsmeldingerForRad,
-        inntektsposter,
-      };
-
-      return acc.concat(ne);
-    },
-    [],
-  );
-
-  const alleInntektsmeldingerSomManglerArbeidsforhold = inntektsmeldinger
-    .filter(im => !arbeidsforhold.some(af => im.arbeidsgiverIdent === af.arbeidsgiverIdent))
-    .map<ArbeidsforholdOgInntektRadData>(im => {
-      const arbeidsgiverOpplysninger = lagArbeidsgiver(
-        im.arbeidsgiverIdent,
-        arbeidsgiverOpplysningerPerId[im.arbeidsgiverIdent],
-      );
-      const inntektsposter = inntekter.find(harMatchendeArbeidsgiverIdent(im))?.inntekter ?? [];
-
-      return {
-        ...arbeidsgiverOpplysninger,
-        årsak: im.årsak ?? undefined,
-        avklaring: lagAvklaringFraInntektsmelding(im),
-        arbeidsforholdForRad: [],
-        inntektsmeldingerForRad: [im],
-        inntektsposter,
-      };
-    });
-
-  return alleArbeidsforhold.concat(alleInntektsmeldingerSomManglerArbeidsforhold).sort(sorterTabell);
-};
-
-const finnUløstArbeidsforholdIndex = (tabellData: ArbeidsforholdOgInntektRadData[]): number[] => {
-  const index = tabellData.findIndex(d => d.årsak && !d.avklaring);
-  return index !== -1 ? [index] : [];
-};
 
 interface Props {
   arbeidOgInntekt: ArbeidOgInntektsmelding;
